set up Ben's code so it's deployable
